Rename DaemonTest to ThreadTest, fix formatting, and update CHANGELOG.md
